requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
kafka-python==2.0.2
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tzdata==2025.1
